tests finally work!
Solution: all UI-independent code now goes in the
"src-model" directory.  UI-dependent code (i.e.,
that can only run in an actual browser) goes
in "src".

The cljsbuild :test target has only "src-model"
and "test" on its :source-paths, so it can run
in phantomjs.

This was quite difficult to get working!